+++ packages/react-query/src/index.ts
@@ -1,17 +1,12 @@
+export { queryOptions } from './queryOptions'
+export { SuspenseQuery } from './SuspenseQuery'
+export { SuspenseInfiniteQuery } from './SuspenseInfiniteQuery'
+
 export { useSuspenseQuery } from './useSuspenseQuery'
-export type {
-  BaseUseSuspenseQueryResult,
-  UseSuspenseQueryOptions,
-  UseSuspenseQueryResultOnLoading,
-  UseSuspenseQueryResultOnSuccess,
-} from './useSuspenseQuery'
+export type { UseSuspenseQueryOptions, UseSuspenseQueryResult } from './useSuspenseQuery'
 export { useSuspenseQueries } from './useSuspenseQueries'
+export type { SuspenseQueriesOptions, SuspenseQueriesResults } from './useSuspenseQueries'
 export { useSuspenseInfiniteQuery } from './useSuspenseInfiniteQuery'
-export type {
-  BaseUseSuspenseInfiniteQueryResult,
-  UseSuspenseInfiniteQueryOptions,
-  UseSuspenseInfiniteQueryResultOnLoading,
-  UseSuspenseInfiniteQueryResultOnSuccess,
-} from './useSuspenseInfiniteQuery'
-export { QueryAsyncBoundary } from './QueryAsyncBoundary'
+export type { UseSuspenseInfiniteQueryOptions, UseSuspenseInfiniteQueryResult } from './useSuspenseInfiniteQuery'
+
 export { QueryErrorBoundary } from './QueryErrorBoundary'