+++ packages/react-query/src/useSuspenseQuery.ts
@@ -1,184 +1,36 @@
-import {
-  type QueryFunction,
-  type QueryKey,
-  type UseQueryOptions,
-  type UseQueryResult,
-  parseQueryArgs,
-  useQuery,
-} from '@tanstack/react-query'
+import { type QueryKey, type UseQueryOptions, type UseQueryResult, useQuery } from '@tanstack/react-query'
 import type { OmitKeyof } from './utility-types'
 
-export interface BaseUseSuspenseQueryResult<TData = unknown>
-  extends OmitKeyof<UseQueryResult<TData>, 'error' | 'isError' | 'isFetching'> {
-  status: 'success' | 'loading'
-}
-
-export interface UseSuspenseQueryResultOnSuccess<TData> extends BaseUseSuspenseQueryResult<TData> {
+export interface UseSuspenseQueryResult<TData = unknown, TError = unknown>
+  extends OmitKeyof<UseQueryResult<TData, TError>, keyof Pick<UseQueryResult, 'isPlaceholderData'>> {
   data: TData
-  isLoading: false
-  isSuccess: true
   status: 'success'
 }
-export interface UseSuspenseQueryResultOnLoading extends BaseUseSuspenseQueryResult {
-  data: undefined
-  isLoading: true
-  isSuccess: false
-  status: 'loading'
-}
 
-export type UseSuspenseQueryOptions<
+export interface UseSuspenseQueryOptions<
   TQueryFnData = unknown,
   TError = unknown,
   TData = TQueryFnData,
   TQueryKey extends QueryKey = QueryKey,
-> = OmitKeyof<UseQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'suspense'>
+> extends OmitKeyof<
+    UseQueryOptions<TQueryFnData, TError, TData, TQueryKey>,
+    'suspense' | 'useErrorBoundary' | 'enabled' | 'placeholderData'
+  > {}
 
 /**
- * This hook wrapping useQuery of @tanstack/react-query with default suspense option. with this hook, you don't have to make unnecessary type narrowing
+ * This hook is wrapping useQuery of `@tanstack/react-query` v4 with default suspense option.
  * @see {@link https://suspensive.org/docs/react-query/useSuspenseQuery}
  */
-// arg1: queryKey, arg2: queryFn, arg3: options
 export function useSuspenseQuery<
   TQueryFnData = unknown,
   TError = unknown,
   TData = TQueryFnData,
   TQueryKey extends QueryKey = QueryKey,
->(
-  queryKey: TQueryKey,
-  queryFn: QueryFunction<TQueryFnData, TQueryKey>,
-  options?: OmitKeyof<
-    UseSuspenseQueryOptions<TQueryFnData, TError, TData, TQueryKey>,
-    'enabled' | 'queryKey' | 'queryFn'
-  >
-): UseSuspenseQueryResultOnSuccess<TData>
-export function useSuspenseQuery<
-  TQueryFnData = unknown,
-  TError = unknown,
-  TData = TQueryFnData,
-  TQueryKey extends QueryKey = QueryKey,
->(
-  queryKey: TQueryKey,
-  queryFn: QueryFunction<TQueryFnData, TQueryKey>,
-  options: OmitKeyof<
-    UseSuspenseQueryOptions<TQueryFnData, TError, TData, TQueryKey>,
-    'enabled' | 'queryKey' | 'queryFn'
-  > & {
-    enabled?: true
-  }
-): UseSuspenseQueryResultOnSuccess<TData>
-export function useSuspenseQuery<
-  TQueryFnData = unknown,
-  TError = unknown,
-  TData = TQueryFnData,
-  TQueryKey extends QueryKey = QueryKey,
->(
-  queryKey: TQueryKey,
-  queryFn: QueryFunction<TQueryFnData, TQueryKey>,
-  options: OmitKeyof<
-    UseSuspenseQueryOptions<TQueryFnData, TError, TData, TQueryKey>,
-    'enabled' | 'queryKey' | 'queryFn'
-  > & {
-    enabled: false
-  }
-): UseSuspenseQueryResultOnLoading
-export function useSuspenseQuery<
-  TQueryFnData = unknown,
-  TError = unknown,
-  TData = TQueryFnData,
-  TQueryKey extends QueryKey = QueryKey,
->(
-  queryKey: TQueryKey,
-  queryFn: QueryFunction<TQueryFnData, TQueryKey>,
-  options: OmitKeyof<UseSuspenseQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'queryKey' | 'queryFn'>
-): UseSuspenseQueryResultOnSuccess<TData> | UseSuspenseQueryResultOnLoading
-
-// arg1: queryKey, arg2: options
-export function useSuspenseQuery<
-  TQueryFnData = unknown,
-  TError = unknown,
-  TData = TQueryFnData,
-  TQueryKey extends QueryKey = QueryKey,
->(
-  queryKey: TQueryKey,
-  options?: OmitKeyof<UseSuspenseQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'enabled' | 'queryKey'>
-): UseSuspenseQueryResultOnSuccess<TData>
-export function useSuspenseQuery<
-  TQueryFnData = unknown,
-  TError = unknown,
-  TData = TQueryFnData,
-  TQueryKey extends QueryKey = QueryKey,
->(
-  queryKey: TQueryKey,
-  options: OmitKeyof<UseSuspenseQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'enabled' | 'queryKey'> & {
-    enabled?: true
-  }
-): UseSuspenseQueryResultOnSuccess<TData>
-export function useSuspenseQuery<
-  TQueryFnData = unknown,
-  TError = unknown,
-  TData = TQueryFnData,
-  TQueryKey extends QueryKey = QueryKey,
->(
-  queryKey: TQueryKey,
-  options: OmitKeyof<UseSuspenseQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'enabled' | 'queryKey'> & {
-    enabled: false
-  }
-): UseSuspenseQueryResultOnLoading
-export function useSuspenseQuery<
-  TQueryFnData = unknown,
-  TError = unknown,
-  TData = TQueryFnData,
-  TQueryKey extends QueryKey = QueryKey,
->(
-  queryKey: TQueryKey,
-  options: OmitKeyof<UseSuspenseQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'queryKey'>
-): UseSuspenseQueryResultOnSuccess<TData> | UseSuspenseQueryResultOnLoading
-
-// arg1: options
-export function useSuspenseQuery<
-  TQueryFnData = unknown,
-  TError = unknown,
-  TData = TQueryFnData,
-  TQueryKey extends QueryKey = QueryKey,
->(
-  options: OmitKeyof<UseSuspenseQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'enabled'> & {
-    enabled?: true
-  }
-): UseSuspenseQueryResultOnSuccess<TData>
-export function useSuspenseQuery<
-  TQueryFnData = unknown,
-  TError = unknown,
-  TData = TQueryFnData,
-  TQueryKey extends QueryKey = QueryKey,
->(
-  options: OmitKeyof<UseSuspenseQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'enabled'> & {
-    enabled: false
-  }
-): UseSuspenseQueryResultOnLoading
-export function useSuspenseQuery<
-  TQueryFnData = unknown,
-  TError = unknown,
-  TData = TQueryFnData,
-  TQueryKey extends QueryKey = QueryKey,
->(
-  options: UseSuspenseQueryOptions<TQueryFnData, TError, TData, TQueryKey>
-): UseSuspenseQueryResultOnSuccess<TData> | UseSuspenseQueryResultOnLoading
-
-// base useSuspenseQuery
-export function useSuspenseQuery<
-  TQueryFnData = unknown,
-  TError = unknown,
-  TData = TQueryFnData,
-  TQueryKey extends QueryKey = QueryKey,
->(
-  arg1: TQueryKey | UseSuspenseQueryOptions<TQueryFnData, TError, TData, TQueryKey>,
-  arg2?:
-    | QueryFunction<TQueryFnData, TQueryKey>
-    | OmitKeyof<UseSuspenseQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'queryKey'>,
-  arg3?: OmitKeyof<UseSuspenseQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'queryKey' | 'queryFn'>
-) {
-  return useQuery({
-    ...parseQueryArgs(arg1, arg2, arg3),
+>(options: UseSuspenseQueryOptions<TQueryFnData, TError, TData, TQueryKey>) {
+  return useQuery<TQueryFnData, TError, TData, TQueryKey>({
+    ...options,
+    enabled: true,
+    useErrorBoundary: true,
     suspense: true,
-  }) as BaseUseSuspenseQueryResult<TData>
+  }) as UseSuspenseQueryResult<TData, TError>
 }