+++ packages/react/src/Suspensive.tsx
@@ -2,24 +2,24 @@ import { type ContextType, type PropsWithChildren, useMemo } from 'react'
 import { DelayDefaultPropsContext, DevModeContext, SuspenseDefaultPropsContext, SuspensiveDevMode } from './contexts'
 import {
   SuspensiveError,
-  Suspensive_config_defaultOptions_delay_ms_should_be_greater_than_0,
+  Suspensive_config_defaultProps_delay_ms_should_be_greater_than_0,
 } from './models/SuspensiveError'
 
 export class Suspensive {
-  public defaultOptions?: {
+  public defaultProps?: {
     suspense?: ContextType<typeof SuspenseDefaultPropsContext>
     delay?: ContextType<typeof DelayDefaultPropsContext>
   }
   public devMode = new SuspensiveDevMode()
 
-  constructor(config: { defaultOptions?: Suspensive['defaultOptions'] } = {}) {
-    if (process.env.NODE_ENV === 'development' && typeof config.defaultOptions?.delay?.ms === 'number') {
+  constructor(config: { defaultProps?: Suspensive['defaultProps'] } = {}) {
+    if (process.env.NODE_ENV === 'development' && typeof config.defaultProps?.delay?.ms === 'number') {
       SuspensiveError.assert(
-        config.defaultOptions.delay.ms > 0,
-        Suspensive_config_defaultOptions_delay_ms_should_be_greater_than_0
+        config.defaultProps.delay.ms > 0,
+        Suspensive_config_defaultProps_delay_ms_should_be_greater_than_0
       )
     }
-    this.defaultOptions = config.defaultOptions
+    this.defaultProps = config.defaultProps
   }
 }
 
@@ -27,13 +27,13 @@ interface SuspensiveProviderProps extends PropsWithChildren {
   value: Suspensive
 }
 export const SuspensiveProvider = ({ value, children }: SuspensiveProviderProps) => {
-  const delayDefaultOptions = useMemo(() => value.defaultOptions?.delay || {}, [value.defaultOptions?.delay])
-  const suspenseDefaultOptions = useMemo(() => value.defaultOptions?.suspense || {}, [value.defaultOptions?.suspense])
+  const delayDefaultProps = useMemo(() => value.defaultProps?.delay || {}, [value.defaultProps?.delay])
+  const suspenseDefaultProps = useMemo(() => value.defaultProps?.suspense || {}, [value.defaultProps?.suspense])
 
   return (
     <DevModeContext.Provider value={value.devMode}>
-      <DelayDefaultPropsContext.Provider value={delayDefaultOptions}>
-        <SuspenseDefaultPropsContext.Provider value={suspenseDefaultOptions}>
+      <DelayDefaultPropsContext.Provider value={delayDefaultProps}>
+        <SuspenseDefaultPropsContext.Provider value={suspenseDefaultProps}>
           {children}
         </SuspenseDefaultPropsContext.Provider>
       </DelayDefaultPropsContext.Provider>